requirements.txt
# Python package requirements
# Add your project dependencies here
# Example:
# numpy==1.24.3
# pandas==2.0.2
torch
opencv-python
pandas
PyYAML
tqdm
matplotlib
seaborn
requests
pyttsx3
ultralytics 